backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.0
python-dotenv==1.0.0
pydantic==2.4.2
langchain==0.1.0
langchain-community==0.0.10
textblob==0.17.1
gotrue==2.4.2
httpx==0.24.1
python-multipart==0.0.6

# Phase 1: Authentication & Security
python-jose[cryptography]==3.3.0  # JWT handling
passlib[bcrypt]==1.7.4            # Password hashing
slowapi==0.1.9                    # Rate limiting
sentry-sdk[fastapi]==1.39.1       # Error tracking
pydantic-settings==2.1.0          # Settings management
email-validator==2.1.0            # Email validation

# Phase 4: AI Features - Using Ollama (local LLM)
# httpx already included above for API calls
